+++ java_public_api/tech/v3/dataset/Reductions.java
@@ -27,7 +27,7 @@ public class Reductions {
   static final IFn probCdfsFn = requiringResolve("tech.v3.dataset.reductions.apache-data-sketch", "prob-cdfs");
   static final IFn probPmfsFn = requiringResolve("tech.v3.dataset.reductions.apache-data-sketch", "prob-pmfs");
   static final IFn probIQRangeFn = requiringResolve("tech.v3.dataset.reductions.apache-data-sketch", "prob-interquartile-range");
-  static final IFn groupByColumnAggFn = requiringResolve("tech.v3.dataset.reductions", "group-by-columns-agg");
+  static final IFn groupByColumnAggFn = requiringResolve("tech.v3.dataset.reductions", "group-by-column-agg");
 
 
   /**
@@ -51,6 +51,54 @@ public class Reductions {
    *    to boolean.  Only indexes for which the index-filter returns true will be added to the
    *    aggregation.  For very large datasets, this is a bit faster than using filter before
    *    the aggregation.
+   * 
+   * Example:
+   *
+   *```java
+   * //Begin parallelized expansion
+   *Iterable dsSeq = (Iterable)rowMapcat(srcds, tallyDays, hashmap(kw("result-type"), kw("as-seq")));
+   *
+   * //The first aggregation is to summarize by placement and simulation the year-month tallies.
+   * //We are essentially replacing count with a summarized count.  After this statement
+   * //we can guarantee that the dataset has unique tuples of [simulation, placement, year-month]
+   *Map initAgg = Reductions.groupByColumnsAgg(dsSeq, vector("simulation", "placement", "year-month"),
+   *					       hashmap("count", Reductions.sum("count")),
+   *					       null);
+   *println(head(initAgg));
+   * //["simulation" "placement" "year-month"]-aggregation [5 4]:
+   *
+   * //| simulation | placement | year-month | count |
+   * //|-----------:|----------:|------------|------:|
+   * //|          0 |         0 |    2020-12 | 622.0 |
+   * //|          0 |         1 |    2020-12 | 591.0 |
+   * //|          0 |         2 |    2020-12 | 500.0 |
+   * //|          0 |         3 |    2020-12 | 549.0 |
+   * //|          0 |         4 |    2020-12 | 595.0 |
+   *
+   * // The second aggregation allows us to build of statistics over each placement/year-month
+   * // pair thus finding out the distribution of a given placement, year-month across simluations
+   *Map result = Reductions.groupByColumnsAgg(vector(initAgg), vector("placement", "year-month"),
+   *					      hashmap("min-count",     Reductions.probQuantile("count", 0.0),
+   *						      "low-95-count",  Reductions.probQuantile("count", 0.05),
+   *						      "q1-count",      Reductions.probQuantile("count", 0.25),
+   *						      "median-count",  Reductions.probQuantile("count", 0.5),
+   *						      "q3-count",      Reductions.probQuantile("count", 0.75),
+   *						      "high-95-count", Reductions.probQuantile("count", 0.95),
+   *						      "max-count",     Reductions.probQuantile("count", 1.0),
+   *						      "count",         Reductions.sum("count")),
+   *					      null);
+   * //Take a million row dataset, expand it, then perform two grouping aggregations.
+   *println(head(result));
+   * //["placement" "year-month"]-aggregation [5 10]:
+   *
+   * //| q3-count | median-count | min-count | high-95-count | placement | max-count |   count | low-95-count | q1-count | year-month |
+   * //|---------:|-------------:|----------:|--------------:|----------:|----------:|--------:|-------------:|---------:|------------|
+   * //|    646.0 |        593.0 |     366.0 |         716.0 |        36 |     809.0 | 58920.0 |        475.0 |    536.0 |    2020-12 |
+   * //|    621.0 |        560.0 |     376.0 |         739.0 |        36 |     782.0 | 57107.0 |        459.0 |    512.0 |    2020-10 |
+   * //|    168.0 |        139.0 |      25.0 |         211.0 |         0 |     246.0 | 13875.0 |         76.0 |    112.0 |    2021-01 |
+   * //|    658.0 |        607.0 |     384.0 |         745.0 |         0 |     825.0 | 60848.0 |        486.0 |    561.0 |    2020-12 |
+   * //|    628.0 |        581.0 |     422.0 |         693.0 |         0 |     802.0 | 58148.0 |        468.0 |    539.0 |    2020-11 |
+   *```
    */
   public static Map groupByColumnsAgg(Iterable dsSeq, Object colname, Map aggMap, Map options) {
     return (Map)groupByColumnAggFn.invoke(colname, aggMap, options, dsSeq);
@@ -217,9 +265,10 @@ public class Reductions {
   /**
    * Probabilistic quantile estimation using default k of 128.
    * See [DoublesSketch](https://datasketches.apache.org/api/java/snapshot/apidocs/index.html).
-   * Multiple quantiles will be merged into a single quantile calculation so it may be more
-   * convenient to use this function to produce multiple quantiles mapped to several result
-   * columns as opposed to ending up with a single column of maps of quantile to value.
+   * Multiple quantile calculations on a single source column will be merged into a single quantile 
+   * calculation so it may be more convenient to use this function to produce multiple quantiles 
+   * mapped to several result columns as opposed to ending up with a single column of maps of quantile 
+   * to value.
    *
    * @param quantile Number from 0-1.
    * @param k Defaults to 128. This produces a normalized rank error of about 1.7%